Fixes small issues for clean install and uninstall

+++ install.sh
@@ -97,11 +97,11 @@ echo ""
 # copy files as it will be when integrated into Venus OS | start
 echo "Copy files and creating symlinks..."
 cp -f /data/venus-os_TailscaleGX/tailscale.combined /usr/bin/tailscale.combined
-if [ ! -L "/usr/bin/tailscale" ]; then
+if [ -e "/usr/bin/tailscale" ] && [ ! -L "/usr/bin/tailscale" ]; then
     rm /usr/bin/tailscale
     ln -s /usr/bin/tailscale.combined /usr/bin/tailscale
 fi
-if [ ! -L "/usr/bin/tailscaled" ]; then
+if [ -e "/usr/bin/tailscaled" ] && [ ! -L "/usr/bin/tailscaled" ]; then
     rm /usr/bin/tailscaled
     ln -s /usr/bin/tailscale.combined /usr/bin/tailscaled
 fi
